[0.3.5] Bug-fix :: API 관리자 권한 확인

+++ E_Global_Zone/app/Http/Controllers/DepartmentController.php
@@ -35,7 +35,8 @@ class DepartmentController extends Controller
     public function store(Request $request): JsonResponse
     {
         $rules = [
-            'dept_name' => 'required|string|unique:departments,dept_name'
+            'dept_name' => 'required|string|unique:departments,dept_name',
+            'guard' => 'required|string|in:admin'
         ];
 
         $validated_result = self::request_validator(
@@ -61,6 +62,7 @@ class DepartmentController extends Controller
     {
         $rules = [
             'dept_name' => 'required|string|unique:departments,dept_name',
+            'guard' => 'required|string|in:admin'
         ];
 
         $validated_result = self::request_validator(
@@ -81,8 +83,16 @@ class DepartmentController extends Controller
      * @param Department $dept_id
      * @return JsonResponse
      */
-    public function destroy(Department $dept_id): JsonResponse
+    public function destroy(Request $request, Department $dept_id): JsonResponse
     {
+        // <<-- Request 요청 관리자 권한 검사.
+        $is_admin = self::is_admin($request);
+
+        if (is_object($is_admin)) {
+            return $is_admin;
+        }
+        // -->>
+
         return $this->department->destroy_department($dept_id);
     }
 }
